Bz(iota) --> Bz(r) and gridded subplots
Restructured Runge-kutta section to be in terms of r and not iota. Made
plots gridded (all one window), and put all actions into specific
functions (makePlots() and pressureAsymptotes()). Code is not stable
whenever iota(r) is not proportional to r, which means it's harder to
make a physically realistic system...